+++ source/ProgArgs.h
@@ -165,6 +165,7 @@ namespace bpt = boost::property_tree;
 #define ARG_S3ENDPOINTS_LONG             "s3endpoints"
 #define ARG_S3FASTGET_LONG               "s3fastget"
 #define ARG_S3FASTPUT_LONG               "s3fastput"
+#define ARG_S3UNSIGNED_LONG              "s3unsigned"
 #define ARG_S3IGNOREERRORS_LONG          "s3ignoreerrors"
 #define ARG_S3LISTOBJ_LONG               "s3listobj"
 #define ARG_S3LISTOBJPARALLEL_LONG       "s3listobjpar"